# feat 修改md文件，修复部分数据读入DB的BUG

+++ douyinService/repository/favoriteRepository.go
@@ -22,6 +22,7 @@ func (fr *FavoriteRepository) GetFavoriteByUserIdAndVideoId(userId uint64, video
 		return false, model.Favorite{}
 	}
 }
+
 func (fr *FavoriteRepository) AddFavorite(favorite model.Favorite) bool {
 	if err := global.DB.Debug().Clauses(clause.OnConflict{
 		Columns:   []clause.Column{{Name: "user_id"}, {Name: "video_id"}},
@@ -30,25 +31,8 @@ func (fr *FavoriteRepository) AddFavorite(favorite model.Favorite) bool {
 		return false
 	}
 	return true
-
-	//var favoriteInfo model.Favorite
-	//result := global.DB.Model(model.Favorite{}).Where("user_id =? and video_id =?", favorite.UserId, favorite.VideoId).First(&favoriteInfo)
-	//if result.Error != nil {
-	//	log.Println(result.Error)
-	//	if result.Error == gorm.ErrRecordNotFound {
-	//		result := global.DB.Model(model.Favorite{}).Create(&favorite)
-	//		if result.Error != nil {
-	//			return false
-	//		}
-	//	}
-	//} else {
-	//	result := global.DB.Debug().Model(model.Favorite{}).Where("user_id=? and video_id=?", favorite.UserId, favorite.VideoId).Update("is_deleted", favorite.IsDeleted)
-	//	if result.Error != nil {
-	//		return false
-	//	}
-	//}
-	//return true
 }
+
 func (fr *FavoriteRepository) DeleteFavoriteById(where interface{}) bool {
 	var favorite model.Favorite
 	if err := fr.Base.DeleteByID(favorite, where); err != nil {
@@ -58,9 +42,7 @@ func (fr *FavoriteRepository) DeleteFavoriteById(where interface{}) bool {
 }
 
 // SaveFavoriteCount 根据VideoId查询对应Video点赞数量，并存入数据库
-func (fr *FavoriteRepository) SaveFavoriteCount(videoId uint64, ans int64) error {
-	//id, err := strconv.ParseInt(videoId, 10, 64)
-	//ans, err := favoriteService.RedisGetVideoFavoriteCount(id)
+func (fr *FavoriteRepository) SaveFavoriteCount(videoId uint64, ans uint32) error {
 	global.DB.Model(model.Video{}).Where("video_id = ?", videoId).Update("favorite_count", ans)
 	return nil
 }
